Refactor walkdir, StatStruct, and change_uripath
Add internal constructor to StatStruct to only save the file name of
symlinks. Additionally, save the root path of a StatStruct Path. For
links, additionally save the target in the form root -> target.
Refactor walkdir to close the channel for non-dir path objects and allow
skipping errors because of denied access.
Refactor handling of trailing slashes in change_uripath.
Add pwd method for URI objects.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>

Refactor walkdir, StatStruct, and change_uripath

Add internal constructor to StatStruct to only save the file name of
symlinks. Additionally, save the root path of a StatStruct Path. For
links, additionally save the target in the form root -> target.
Refactor walkdir to close the channel for non-dir path objects and allow
skipping errors because of denied access.
Refactor handling of trailing slashes in change_uripath.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>