Refactor walkdir, rm, cd, splitdir, add basename
Ensure, cd works only for directories and links to directories and
throws and error for non-existent dirs or non-directories (e.g. files).
Update rm, to ensure directories are only removed when empty or the
recursive flag is set. Refactor walkdir to ensure a closed channel is
return, if the initial root is not a directory or link to a directory.
Refactor splitdir and add new methods for basename of URIs and Clients.
Update helper functions for link targets.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>