Add path validity checks, bug fixes, refactoring
Fix analyse_path to consider the root. Only return the flag for dirs,
stop returning the dir name.
Throw an IOError, if pwd does not return a valid path on the Client.
Add a path check for readdir to ensure an IOError is thrown for
non-existant paths.
Add function mkdir similar to Julia's mkdir.
Update docstrings.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>